renamed recording page for clarity
